+++ teleoscope.ca/src/components/landing/demoCTA.tsx
@@ -4,7 +4,7 @@ import { Button } from "../ui/button";
 export default function DemoCTA() {
   return (
     <div className="flex justify-center items-center w-full gap-8 p-10  ">
-      <div className="flex flex-col gap-8 bg-pink-50 max-w-4xl rounded-xl w-full p-10 px-20">
+      <div className="flex flex-col gap-8 bg-neutral-50 shadow max-w-4xl rounded-xl w-full p-10 px-20">
         <div className="text-xl ">
           <p>
             Ready to see Teleoscope first hand? Schedule a demo with our team
