+++ teleoscope.ca/src/components/landing/footer.tsx
@@ -1,8 +1,8 @@
 
-
+// TODO: Add footer content
 export default function Footer() {
     return (
-        <footer className="flex flex-col items-center justify-center gap-4 bg-gray-50 h-24">
+        <footer className="flex flex-col items-center justify-center gap-4 bg-neutral-50 h-24">
             <span className="text-2xl font-bold">Teleoscope</span>
         </footer>
     );
